Added bank selection flow implementation

+++ app/src/main/java/com/example/luismunoz/paymentapp/ui/fragment/AmountEnterFragment.kt
@@ -60,7 +60,7 @@ class AmountEnterFragment : Fragment() {
                         binding.tilAmountEnterFragmentAmountContainer.error = null
 
                         val amountValue = binding.edAmountEnterFragmentValue.text.toString()
-                        val action = AmountEnterFragmentDirections.actionAmountEnterFragmentToPaymentSelectionFragment(amount = amountValue.toLong())
+                        val action = AmountEnterFragmentDirections.actionAmountEnterFragmentToPaymentSelectionFragment(amount = amountValue)
                         findNavController().navigate(action)
                     }
                 }
